Keep Gestures with SpotlightSlider
