+++ appyx-components/spotlight/common/src/commonMain/kotlin/com/bumble/appyx/components/spotlight/ui/slider/SpotlightSlider.kt
@@ -1,6 +1,8 @@
 package com.bumble.appyx.components.spotlight.ui.slider
 
 import androidx.compose.foundation.gestures.Orientation
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
@@ -8,7 +10,15 @@ import com.bumble.appyx.components.spotlight.SpotlightModel.State
 import com.bumble.appyx.components.spotlight.SpotlightModel.State.ElementState.CREATED
 import com.bumble.appyx.components.spotlight.SpotlightModel.State.ElementState.DESTROYED
 import com.bumble.appyx.components.spotlight.SpotlightModel.State.ElementState.STANDARD
+import com.bumble.appyx.components.spotlight.operation.Next
+import com.bumble.appyx.components.spotlight.operation.Previous
+import com.bumble.appyx.interactions.core.ui.context.TransitionBounds
 import com.bumble.appyx.interactions.core.ui.context.UiContext
+import com.bumble.appyx.interactions.core.ui.gesture.Drag
+import com.bumble.appyx.interactions.core.ui.gesture.Gesture
+import com.bumble.appyx.interactions.core.ui.gesture.GestureFactory
+import com.bumble.appyx.interactions.core.ui.gesture.dragHorizontalDirection
+import com.bumble.appyx.interactions.core.ui.gesture.dragVerticalDirection
 import com.bumble.appyx.interactions.core.ui.property.impl.Alpha
 import com.bumble.appyx.interactions.core.ui.property.impl.GenericFloatProperty
 import com.bumble.appyx.interactions.core.ui.property.impl.Position
@@ -76,4 +86,48 @@ class SpotlightSlider<InteractionTarget : Any>(
     ): MutableUiState =
         targetUiState.toMutableState(uiContext, scrollX.renderValueFlow, width)
 
+
+    class Gestures<InteractionTarget>(
+        transitionBounds: TransitionBounds,
+        private val orientation: Orientation = Orientation.Horizontal,
+        private val reverseOrientation: Boolean = false,
+    ) : GestureFactory<InteractionTarget, State<InteractionTarget>> {
+        private val width = transitionBounds.widthPx.toFloat()
+        private val height = transitionBounds.heightPx.toFloat()
+
+        override fun createGesture(
+            state: State<InteractionTarget>,
+            delta: Offset,
+            density: Density
+        ): Gesture<InteractionTarget, State<InteractionTarget>> = when (orientation) {
+            Orientation.Horizontal -> {
+                when (dragHorizontalDirection(delta)) {
+                    Drag.HorizontalDirection.LEFT -> Gesture(
+                        operation = if (reverseOrientation) Previous() else Next(),
+                        completeAt = Offset(-width, 0f)
+                    )
+
+                    else -> Gesture(
+                        operation = if (reverseOrientation) Next() else Previous(),
+                        completeAt = Offset(width, 0f)
+                    )
+                }
+            }
+
+            Orientation.Vertical -> {
+                when (dragVerticalDirection(delta)) {
+                    Drag.VerticalDirection.UP -> Gesture(
+                        operation = if (reverseOrientation) Previous() else Next(),
+                        completeAt = Offset(0f, -height)
+                    )
+
+                    else ->
+                        Gesture(
+                            operation = if (reverseOrientation) Next() else Previous(),
+                            completeAt = Offset(0f, height)
+                        )
+                }
+            }
+        }
+    }
 }
