+++ demos/appyx-interactions/android/src/main/kotlin/com/bumble/appyx/interactions/sample/SpotlightExperimentDebug.kt
@@ -13,7 +13,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.bumble.appyx.components.spotlight.Spotlight
 import com.bumble.appyx.components.spotlight.SpotlightModel
-import com.bumble.appyx.components.spotlight.gestures.Gestures
 import com.bumble.appyx.components.spotlight.operation.first
 import com.bumble.appyx.components.spotlight.operation.last
 import com.bumble.appyx.components.spotlight.operation.next
@@ -42,7 +41,7 @@ fun SpotlightExperimentDebug(modifier: Modifier = Modifier) {
                 savedStateMap = null
             ),
             motionController = { SpotlightSlider(it) },
-            gestureFactory = { Gestures(it) },
+            gestureFactory = { SpotlightSlider.Gestures(it) },
             isDebug = true
         )
     }
